requirements.txt
Flask
google-generativeai
python-dotenv
ultralytics
segmentation-models-pytorch
timm
nibabel
pydicom
pylibjpeg
seaborn
tqdm
scikit-learn